+++ set3/17/sum1.cc
@@ -4,7 +4,7 @@ using namespace std;
 int sum(int argc, char *argv[], int)   // sum1.cc
 {
     int total = 0;
-    for (int idx = 1; idx < argc; ++idx)
+    for (size_t idx = 1; idx < argc; ++idx)
     {
         total += stoi(argv[idx]);
     }
