+++ set3/17/sum2.cc
@@ -4,7 +4,7 @@ using namespace std;
 double sum(int argc, char *argv[], double)   // 2.cc
 {
     double total = 0.0;
-    for (int idx = 1; idx < argc; ++idx)
+    for (size_t idx = 1; idx < argc; ++idx)
     {
         total += stod(argv[idx]);
     }
